Migrated from session to entitymanager. Implemented all DAO methods for Certificate and Tag entities. Covered all DAO methods with integration tests using in-memory database hsqldb with 100%.

+++ repository/src/main/java/com/chupin/DAO/TagDAO.java
@@ -8,9 +8,9 @@ public interface TagDAO {
 
     List<Tag> getAll();
 
-    Tag getById(int id);
+    Tag getById(long id);
 
-    int save(Tag tag);
+    long save(Tag tag);
 
     void update(Tag tag);
